ASoC: wcd9xxx: Add support for MAD Slim master channel API
LAB feature is used in listen to capture PCM data once
the keyword is detected. To get the PCM from codec Apps
processor will enable the Slim master channel. PCM data
can be acquired from this master channel which is further
connected to MAD Tx port in codec. This change implements
and exports API to enable the Slim master channel on Apps
side.

Change-Id: Ie2e3294fbe54fc38e7a8dc971337ea0eac777b06
Signed-off-by: Venkat Sudhir <vsudhir@codeaurora.org>