Revert "msm8974-oppo: Optimize wakeup delay for JDI panel"
This reverts commit 9401bd2eea0142f2de39ca163f4e61728b607aa8.

Display anomalies have been seen on some JDI panels after
introducing this change, such as the display contents being
offset and wrapping vertically.

Revert it for now. It looks correct, but since the latency
improvement it provides is relatively small and the timing
of the change correlates well with the reports of the issue,
it's the safest move.

Issue-Id: OPO-480
Change-Id: Ie413f707f0951834003c05ce7e4e598bfb618b92
Signed-off-by: Matt Wagantall <mwagantall@cyngn.com>